+++ lib/src/main/java/com/ozan/lib/slidablecounterbutton/SlidableCounterButtonViewState.kt
@@ -7,10 +7,10 @@ import kotlin.math.max
 @Parcelize
 data class SlidableCounterButtonViewState(
     val title: String,
-    val pieceValue: Int,
+    val pieceValue: Double,
     val pieceValueSign: String,
     val availableCount: Int,
     val purchasedCount: Int? = 0
 ) : Parcelable {
-    fun getTotalPrice(): Int = max(purchasedCount.orZero(), 1).times(pieceValue)
+    fun getTotalPrice(): Double = max(purchasedCount.orZero(), 1).times(pieceValue)
 }
